backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
motor
pydantic[email]
python-dotenv